Split skills module into package to match project conventions
