+++ src/pydantic_harness/__init__.py
@@ -7,7 +7,7 @@ Usage:
 # Each capability module is imported and re-exported here.
 # Capabilities are listed alphabetically.
 
-from pydantic_harness.skills import Skill, Skills, load_skills_from_directory
+from pydantic_harness.skills import Skill, Skills, load_skills_from_directory  # noqa: F401
 
 __all__: list[str] = [
     'Skill',
