+++ src/pydantic_harness/skills/__init__.py
@@ -1,0 +1,6 @@
+"""Skills capability for progressive tool loading."""
+
+from pydantic_harness.skills._capability import Skills
+from pydantic_harness.skills._toolset import Skill, load_skills_from_directory
+
+__all__ = ['Skill', 'Skills', 'load_skills_from_directory']
